Use decimal values from EVM scan fee API
We should include the decimal value in our calculation to be more
precise in our conversion from GWEI to WEI. This is important for some
currencies like Fantom which will reject transactions if their fee is
too low which may happen due to rounding down.

+++ src/ethereum/fees/feeProviders.ts
@@ -168,9 +168,9 @@ export const fetchFeesFromEvmScan = async (
 
   const { SafeGasPrice, ProposeGasPrice, FastGasPrice } =
     asEvmScanGasResponseResult(esGasResponse.result)
-  const newSafeLow = parseInt(SafeGasPrice)
-  let newAverage = parseInt(ProposeGasPrice)
-  let newFast = parseInt(FastGasPrice)
+  const newSafeLow = parseFloat(SafeGasPrice)
+  let newAverage = parseFloat(ProposeGasPrice)
+  let newFast = parseFloat(FastGasPrice)
 
   // Correct inconsistencies, convert values
   if (newAverage <= newSafeLow) newAverage = newSafeLow + 1
